+++ lib/aerospike/command/batch_index_command.rb
@@ -62,7 +62,7 @@ module Aerospike
         end
       end
       size_buffer
-      write_header_read(policy, read_attr | INFO1_BATCH, 0, field_count, 0)
+      write_header_read(policy, read_attr | INFO1_BATCH, 0, 0, field_count, 0)
 
       write_filter_exp(@policy.filter_exp, exp_size)