+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+## [4.1.0] 2024-10-22
+
+- **New Features**
+  - [CLIENT-2833] Support `Policy#ReadTouchTtlPercent`.
+  - [CLIENT-2826] Support `QueryDuration` in `QueryPolicy#ExpectedDuration`.
+  - [CLIENT-3103] Support `XDR_KEY_BUSY`.
+
+- **Fixes**
+  - [CLIENT-3144] Various fixes. PR #132 and #133 Thanks to [Igor Pstyga](https://github.com/opti)
+    - Fix `BatchRead` for multiple records with operations.
+    - Use correct namespace for the `MapReturnType`.
+    - `BatchRead` with operations would throw an exception.
+  - Fix a test with invalid map key in Server v7.1.
+
 ## [4.0.0] 2024-08-14
 
 - **New Features**